MULE-11776: Make the WSC use the Soap Service
