+++ extensions/wsc/src/main/java/org/mule/extension/ws/internal/security/SecurityStrategyAdapter.java
@@ -4,7 +4,7 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.extension.ws.api.security;
+package org.mule.extension.ws.internal.security;
 
 import org.mule.runtime.extension.api.annotation.Alias;
 import org.mule.services.soap.api.security.SecurityStrategy;
