+++ extensions/wsc/src/main/java/org/mule/extension/ws/internal/security/WssEncryptSecurityStrategy.java
@@ -4,25 +4,23 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.extension.ws.api.security;
+package org.mule.extension.ws.internal.security;
 
-import org.mule.extension.ws.api.security.config.WssKeyStoreConfigurationAdapter;
+import org.mule.extension.ws.internal.security.config.WssKeyStoreConfigurationAdapter;
 import org.mule.runtime.extension.api.annotation.param.Parameter;
+import org.mule.services.soap.api.security.EncryptSecurityStrategy;
 import org.mule.services.soap.api.security.SecurityStrategy;
-import org.mule.services.soap.api.security.SignSecurityStrategy;
 import org.mule.services.soap.api.security.config.WssKeyStoreConfiguration;
 
 /**
- * Signs the SOAP request that is being sent, using the private key of the key-store in the provided TLS context.
+ * Verifies the signature of a SOAP response, using certificates of the trust-store in the provided TLS context.
  *
  * @since 4.0
  */
-public class WssSignSecurityStrategy implements SecurityStrategyAdapter {
-
-  private static final String WS_SIGN_PROPERTIES_KEY = "signProperties";
+public class WssEncryptSecurityStrategy implements SecurityStrategyAdapter {
 
   /**
-   * The keystore to use when signing the message.
+   * The keystore to use when encrypting the message.
    */
   @Parameter
   private WssKeyStoreConfigurationAdapter keyStoreConfiguration;
@@ -33,6 +31,6 @@ public class WssSignSecurityStrategy implements SecurityStrategyAdapter {
         new WssKeyStoreConfiguration(keyStoreConfiguration.getAlias(), keyStoreConfiguration.getKeyPassword(),
                                      keyStoreConfiguration.getPassword(), keyStoreConfiguration.getStorePath(),
                                      keyStoreConfiguration.getType());
-    return new SignSecurityStrategy(keyStore);
+    return new EncryptSecurityStrategy(keyStore);
   }
 }
