+++ extensions/wsc/src/main/java/org/mule/extension/ws/internal/security/WssUsernameTokenSecurityStrategy.java
@@ -5,7 +5,7 @@
  * LICENSE.txt file.
  */
 
-package org.mule.extension.ws.api.security;
+package org.mule.extension.ws.internal.security;
 
 import org.mule.runtime.extension.api.annotation.param.Optional;
 import org.mule.runtime.extension.api.annotation.param.Parameter;
