+++ extensions/wsc/src/main/java/org/mule/extension/ws/internal/security/WssVerifySignatureSecurityStrategy.java
@@ -4,10 +4,10 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.extension.ws.api.security;
+package org.mule.extension.ws.internal.security;
 
 
-import org.mule.extension.ws.api.security.config.WssTrustStoreConfigurationAdapter;
+import org.mule.extension.ws.internal.security.config.WssTrustStoreConfigurationAdapter;
 import org.mule.runtime.extension.api.annotation.param.Optional;
 import org.mule.runtime.extension.api.annotation.param.Parameter;
 import org.mule.services.soap.api.security.SecurityStrategy;
