+++ services/soap/src/test/java/org/mule/services/soap/runtime/attachments/AttachmentsTestCase.java
@@ -10,7 +10,7 @@ import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.hasSize;
 import static org.junit.Assert.assertThat;
 import static org.mule.services.soap.api.message.SoapRequest.builder;
-import org.mule.services.soap.api.message.ImmutableSoapAttachment;
+import org.mule.services.soap.api.message.SoapAttachment;
 import org.mule.services.soap.api.message.SoapRequest;
 import org.mule.services.soap.api.message.SoapResponse;
 import org.mule.services.soap.AbstractSoapServiceTestCase;
@@ -44,8 +44,8 @@ public class AttachmentsTestCase extends AbstractSoapServiceTestCase {
   @Description("Uploads an attachment to a mtom server")
   public void uploadAttachment() throws Exception {
     SoapRequest request = builder()
-        .withAttachment(new ImmutableSoapAttachment("attachment", "text/html",
-                                                    new ByteArrayInputStream("Some Content".getBytes())))
+        .withAttachment(new SoapAttachment("attachment", "text/html",
+                                           new ByteArrayInputStream("Some Content".getBytes())))
         .withContent("<con:uploadAttachment xmlns:con=\"http://service.soap.services.mule.org/\"/>")
         .withOperation("uploadAttachment")
         .build();
